auto_curation/requirements.txt
streamlit
requests
python-dotenv
google-generativeai
nltk
sqlalchemy
pillow
